rename AttrSerde to JSONModelSerde
